+++ R/flux_calc.R
@@ -15,9 +15,9 @@
 #' \ifelse{html}{\out{micromol * m<sup>-2</sup> * h<sup>-1</sup>}}{\eqn{micromol*m^{-2}*h^{-1}}{ASCII}}
 #' @param f_cut column containing cutting information
 #' @param keep_arg name in `f_cut` of data to keep
-#' @param chamber_volume volume of the flux chamber in L,
-#' can also be a column in case it is a variable
-#' @param tube_volume volume of the tubing in L,
+#' @param chamber_volume `r lifecycle::badge("deprecated")` see `setup_volume`
+#' @param tube_volume `r lifecycle::badge("deprecated")` see `setup_volume`
+#' @param setup_volume volume of the flux chamber and instrument together in L,
 #' can also be a column in case it is a variable
 #' @param atm_pressure atmospheric pressure in atm,
 #' can be a constant (numerical) or a variable (column name)
@@ -36,7 +36,8 @@
 #' for each flux in the output. Note that NA are removed in median calculation.
 #' Those columns will get the `_med` suffix in the output.
 #' @param cols_nest columns to nest in `nested_variables` for each flux in the
-#' output.
+#' output. Can be character vector of column names, `"none"` (default) selects
+#' none, or `"all"` selects all the column except those in `cols_keep`.
 #' @param f_fluxid column containing the flux IDs
 #' @param temp_air_col column containing the air temperature used
 #' to calculate fluxes. Will be averaged with NA removed.
@@ -66,6 +67,7 @@
 #' ungroup mutate case_when distinct left_join across everything
 #' @importFrom tidyselect any_of
 #' @importFrom stats median
+#' @importFrom lifecycle deprecated deprecate_stop
 #' @examples
 #' data(co2_conc)
 #' slopes <- flux_fitting(co2_conc, conc, datetime, fit_type = "exp_zhao18")
@@ -75,8 +77,7 @@
 #' temp_air,
 #' conc_unit = "ppm",
 #' flux_unit = "mmol",
-#' chamber_volume = 24.5,
-#' tube_volume = 0.075,
+#' setup_volume = 24.575,
 #' atm_pressure = 1,
 #' plot_area = 0.0625)
 #' @export
@@ -87,7 +88,8 @@ flux_calc <- function(slopes_df,
                       slope_col,
                       f_datetime = f_datetime,
                       temp_air_col,
-                      chamber_volume,
+                      chamber_volume = deprecated(),
+                      setup_volume,
                       atm_pressure,
                       plot_area,
                       f_fluxid = f_fluxid,
@@ -97,13 +99,29 @@ flux_calc <- function(slopes_df,
                       cols_ave = c(),
                       cols_sum = c(),
                       cols_med = c(),
-                      cols_nest = c(),
-                      tube_volume,
+                      cols_nest = "none",
+                      tube_volume = deprecated(),
                       temp_air_unit = "celsius",
                       f_cut = f_cut,
                       keep_arg = "keep",
                       cut = TRUE,
                       fit_type = c()) {
+
+  if (is_present(chamber_volume)) {
+    deprecate_stop(
+      when = "1.2.2",
+      what = "flux_calc(chamber_volume)",
+      with = "flux_calc(setup_volume)"
+    )
+  }
+
+  if (is_present(tube_volume)) {
+    deprecate_stop(
+      when = "1.2.2",
+      what = "flux_calc(tube_volume)",
+      with = "flux_calc(setup_volume)"
+    )
+  }
 
   name_df <- deparse(substitute(slopes_df))
 
@@ -140,6 +158,18 @@ flux_calc <- function(slopes_df,
 
   if (any(!df_ok))
     stop("Please correct the arguments", call. = FALSE)
+
+  if (length(cols_nest) == 1 && cols_nest == "all") {
+    cols_nest <- slopes_df |>
+      select(!c(
+        {{cols_keep}}
+      )) |>
+      names()
+  }
+
+  if (length(cols_nest) == 1 && cols_nest == "none") {
+    cols_nest <- c()
+  }
 
 
   fit_type <- flux_fit_type(
@@ -183,7 +213,7 @@ flux_calc <- function(slopes_df,
   }
 
 
-  name_vol <- deparse(substitute(chamber_volume))
+  name_vol <- deparse(substitute(setup_volume))
   name_atm <- deparse(substitute(atm_pressure))
   name_plot <- deparse(substitute(plot_area))
 
@@ -280,19 +310,6 @@ flux_calc <- function(slopes_df,
     slope_med <- slope_sum
   }
 
-  # if (length(cols_nest) > 0) {
-  #   message("Creating a df with the columns from 'cols_nest' argument...")
-  #   slope_nest <- slopes_df |>
-  #     select(all_of(cols_nest), {{f_fluxid}}) |>
-  #     nest_by(
-  #       {{f_fluxid}}, .key = "nested_variables"
-  #     ) |>
-  #     left_join(slope_med, by = join_by(
-  #       {{f_fluxid}} == {{f_fluxid}}
-  #     ))
-  # } else {
-  #   slope_nest <- slope_med
-  # }
 
   message("Calculating fluxes...")
 
@@ -315,9 +332,8 @@ flux_calc <- function(slopes_df,
 
   fluxes <- slope_med |>
     mutate(
-      f_volume_setup = {{chamber_volume}} + tube_volume,
       f_flux =
-        ({{slope_col}} * {{atm_pressure}} * .data$f_volume_setup)
+        ({{slope_col}} * {{atm_pressure}} * {{setup_volume}})
         / (r_const *
            .data$f_temp_air_ave
            * {{plot_area}}) # flux in micromol/s/m^2