+++ R/flux_match.R
@@ -9,18 +9,19 @@
 #' @param field_record dataframe recording which measurement happened when.
 #' Has to contain at least a column containing the start of each measurement,
 #' and any other column identifying the measurements.
-#' @param startcrop how many seconds should be discarded at the beginning of
-#' the measurement
+#' @param startcrop `r lifecycle::badge("deprecated")` `startcrop` is no longer
+#' supported. Please use `start_cut` in `flux_fitting` instead.
 #' @param measurement_length length of the measurement (in seconds)
 #' from the start specified in the `field_record`
-#' @param ratio_threshold ratio (number of concentration measurement compared to
-#' length of measurement in seconds) below which the data should be flagged as
-#' too little
+#' @param ratio_threshold `r lifecycle::badge("deprecated")` `ratio_threshold`
+#' is no longer supported. Please use `ratio_threshold` in `flux_quality`
+#' instead.
 #' @param time_diff time difference (in seconds) between the two datasets.
 #' Will be added to the datetime column of the `raw_conc` dataset.
 #' For situations where the time was not synchronized correctly.
 #' @param f_datetime datetime column in raw_conc (`ymd_hms` format)
-#' @param f_conc concentration column in `raw_conc`
+#' @param f_conc `r lifecycle::badge("deprecated")` `f_conc` is no longer
+#' required
 #' @param start_col start column in field_record (`ymd_hms` format)
 #' @param end_col end columne in field_record (`ymd_hms` format).
 #' Only needed if `fixed_length = "FALSE"`.
@@ -36,9 +37,10 @@
 #' group_by filter ungroup select distinct pull join_by coalesce
 #' @importFrom tidyr fill drop_na
 #' @importFrom lubridate is.POSIXct
+#' @importFrom lifecycle deprecate_stop deprecated deprecate_warn is_present
 #' @examples
 #' data(co2_df_short, record_short)
-#' flux_match(co2_df_short, record_short, datetime, start, conc, startcrop = 10,
+#' flux_match(co2_df_short, record_short, datetime, start,
 #' measurement_length = 180)
 #' @export
 
@@ -47,37 +49,56 @@ flux_match <- function(raw_conc,
                        field_record,
                        f_datetime,
                        start_col,
-                       f_conc,
                        end_col,
-                       startcrop,
                        measurement_length,
                        fixed_length = TRUE,
-                       ratio_threshold = 0.5,
-                       time_diff = 0) {
+                       time_diff = 0,
+                       startcrop = 0,
+                       ratio_threshold = deprecated(),
+                       f_conc = deprecated()) {
+
+  if (startcrop != 0) {
+    deprecate_stop(
+      when = "1.2.1",
+      what = "flux_match(startcrop)",
+      with = "flux_fitting(start_cut)"
+    )
+  }
+
+  if (is_present(ratio_threshold)) {
+    deprecate_warn(
+      when = "1.2.2",
+      what = "flux_match(ratio_threshold)",
+      with = "flux_quality(ratio_threshold)"
+    )
+  }
+
+  if (is_present(f_conc)) {
+    deprecate_warn(
+      when = "1.2.2",
+      what = "flux_match(f_conc)",
+      details = "f_conc is no longer required"
+    )
+  }
 
   name_raw_conc <- deparse(substitute(raw_conc))
   name_field_record <- deparse(substitute(field_record))
 
-  args_ok <- flux_fun_check(list(
-    startcrop = startcrop,
-    ratio_threshold = ratio_threshold,
-    time_diff = time_diff
-  ),
-  fn = list(is.numeric, is.numeric, is.numeric),
-  msg = rep("has to be numeric", 3))
+  args_ok <- flux_fun_check(
+    list(time_diff = time_diff),
+    fn = list(is.numeric),
+    msg = "has to be numeric"
+  )
 
   raw_conc_check <- raw_conc |>
-    select({{f_datetime}}, {{f_conc}})
+    select({{f_datetime}})
 
   field_record_check <- field_record |>
     select({{start_col}})
 
   raw_conc_ok <- flux_fun_check(raw_conc_check,
-                                fn = list(is.POSIXct, is.numeric),
-                                msg = c(
-                                  "has to be POSIXct",
-                                  "has to be numeric"
-                                ),
+                                fn = list(is.POSIXct),
+                                msg = "has to be POSIXct",
                                 name_df = name_raw_conc)
 
   field_record_ok <- flux_fun_check(field_record_check,
@@ -91,15 +112,13 @@ flux_match <- function(raw_conc,
 
 
 
-  if (ratio_threshold < 0 || ratio_threshold > 1) {
-    stop("ratio_threshold has to be a number between 0 and 1")
-  }
+
 
 
   field_record <- field_record |>
     arrange({{start_col}}) |>
     mutate(
-      f_start = {{start_col}} + startcrop,
+      f_start = {{start_col}},
       f_fluxid = row_number()
     )
 
@@ -148,44 +167,13 @@ flux_match <- function(raw_conc,
       ({{f_datetime}} < .data$f_end &
          {{f_datetime}} >= .data$f_start)
     ) |>
-    mutate(
-      f_n_conc = sum(!is.na({{f_conc}})),
-      f_length = difftime(.data$f_end, .data$f_start, units = "secs"),
-      f_length = as.numeric(.data$f_length),
-      f_ratio = .data$f_n_conc / .data$f_length,
-      f_flag_match = case_when(
-        .data$f_ratio == 0 ~ "no data",
-        .data$f_ratio <= ratio_threshold ~ "nb of data too low"
-      )
-    ) |>
     ungroup()
 
   conc_df <- conc_df |>
     mutate(
       f_fluxid = as.factor(.data$f_fluxid),
-      f_flag_match = as.character(.data$f_flag_match)
     ) |>
     arrange(.data$f_fluxid)
-
-
-  flags <- conc_df |>
-    select("f_fluxid", "f_flag_match") |>
-    drop_na("f_flag_match") |>
-    distinct() |>
-    mutate(
-      f_warnings = paste(
-        "\n", "fluxID", .data$f_fluxid, ":",
-        .data$f_flag_match
-      ),
-      f_warnings = as.character(.data$f_warnings)
-    ) |>
-    pull(.data$f_warnings)
-
-  f_warnings <- str_c(flags)
-
-
-  if (any(!is.na(conc_df$f_flag_match))) warning(f_warnings)
-
 
   conc_df
 }