+++ tests/testthat/test-flux_match.R
@@ -4,11 +4,9 @@ test_that("matching works", {
     record_short,
     datetime,
     start,
-    conc,
-    startcrop = 10,
     measurement_length = 180
   ) |>
-    dplyr::select(f_fluxid, f_n_conc, f_ratio, f_flag_match, f_length) |>
+    dplyr::select(f_fluxid, f_start, f_end) |>
     dplyr::distinct()
   )
 })
@@ -26,8 +24,6 @@ test_that("time_diff works", {
     record_short,
     datetime,
     start,
-    conc,
-    startcrop = 10,
     measurement_length = 220,
     time_diff = 180
   ))
@@ -53,8 +49,6 @@ test_that("renaming variables works", {
       record_short,
       date_time,
       starting,
-      CO2_conc,
-      startcrop = 10,
       measurement_length = 220
     )
   )
@@ -70,43 +64,14 @@ test_that("flags on nb of data", {
         record_short,
         datetime,
         start,
-        conc,
-        startcrop = 10,
         measurement_length = 220
       )
     )
   )
 })
 
-# test that flags also get printed as warnings
-test_that("warnings", {
-  expect_warning(
-    flux_match(
-      co2_df_missing,
-      record_short,
-      datetime,
-      start,
-      conc,
-      startcrop = 10,
-      measurement_length = 220
-    ),
-    "fluxID 1 : nb of data too low
- fluxID 3 : nb of data too low
- fluxID 6 : no data"
-  )
-})
 
-test_that("no warnings when no flags", {
-  expect_no_warning(flux_match(
-    co2_df_short,
-    record_short,
-    datetime,
-    start,
-    conc,
-    startcrop = 10,
-    measurement_length = 220
-  ))
-})
+
 
 # test that the data type checking works (all the error messages)
 
@@ -122,33 +87,12 @@ test_that("error on datetime", {
       record_short,
       datetime,
       start,
-      conc,
-      startcrop = 10,
       measurement_length = 220
     ),
     "Please correct the arguments"
   )
 })
 
-test_that("error on conc variable", {
-  co2_df_short <- co2_df_short %>%
-    dplyr::mutate(
-      conc = as.character(conc)
-    )
-
-  expect_error(
-    flux_match(
-      co2_df_short,
-      record_short,
-      datetime,
-      start,
-      conc,
-      startcrop = 10,
-      measurement_length = 220
-    ),
-    "Please correct the arguments"
-  )
-})
 
 test_that("error on start", {
   record_short <- record_short %>%
@@ -162,23 +106,6 @@ test_that("error on start", {
       record_short,
       datetime,
       start,
-      conc,
-      startcrop = 10,
-      measurement_length = 220
-    ),
-    "Please correct the arguments"
-  )
-})
-
-test_that("error on startcrop", {
-  expect_error(
-    flux_match(
-      co2_df_short,
-      record_short,
-      datetime,
-      start,
-      conc,
-      startcrop = "blip",
       measurement_length = 220
     ),
     "Please correct the arguments"
@@ -192,29 +119,12 @@ test_that("error on measurement_length", {
       record_short,
       datetime,
       start,
-      conc,
-      measurement_length = "blip",
-      startcrop = 10
+      measurement_length = "blip"
     ),
     "Please correct the arguments"
   )
 })
 
-test_that("error on ratio_threshold", {
-  expect_error(
-    flux_match(
-      co2_df_short,
-      record_short,
-      datetime,
-      start,
-      conc,
-      ratio_threshold = 2,
-      startcrop = 10,
-      measurement_length = 220
-    ),
-    "ratio_threshold has to be a number between 0 and 1"
-  )
-})
 
 test_that("error on time_diff", {
   expect_error(
@@ -223,8 +133,6 @@ test_that("error on time_diff", {
       record_short,
       datetime,
       start,
-      conc,
-      startcrop = 10,
       measurement_length = 220,
       time_diff = "comment est votre blanquette?"
     ),
@@ -247,12 +155,10 @@ test_that("matching works with end col", {
     record_short_end,
     datetime,
     start,
-    conc,
     end,
-    startcrop = 10,
     fixed_length = FALSE
   ) |>
-    dplyr::select(f_fluxid, f_ratio, f_flag_match, f_length) |>
+    dplyr::select(f_fluxid, f_start, f_end) |>
     dplyr::distinct()
   )
 })